refactor: pickeat 관련 도메인 디렉터리 정리
