+++ frontend/src/domains/pickeat/matchResult/components/PickeatDecisionInfo.tsx
