+++ frontend/src/domains/pickeat/matchResult/components/PickeatVoteCompleteButton.tsx
