+++ frontend/src/domains/pickeat/matchResult/components/Result.tsx
