+++ frontend/src/domains/pickeat/participants/participantsAvatarGroup/ParticipantAvatar.tsx
