+++ frontend/src/domains/pickeat/participants/participantsAvatarGroup/ParticipantInfoTooltip.tsx
