+++ frontend/src/domains/pickeat/participants/participantsAvatarGroup/ParticipantsAvatarGroup.tsx
