+++ frontend/src/domains/pickeat/participants/participantsAvatarGroup/hooks/useParticipantsState.ts
