+++ frontend/src/domains/pickeat/preferRestaurant/components/Participant.tsx
