+++ frontend/src/domains/pickeat/restaurantExclude/components/restaurantTabList/restaurantList/ExcludedRestaurantItem.tsx
