+++ frontend/src/domains/pickeat/restaurantExclude/components/restaurantTabList/restaurantList/RestaurantItem.stories.tsx
