+++ frontend/src/domains/pickeat/restaurantExclude/components/restaurantTabList/restaurantList/RestaurantList.stories.tsx
