+++ frontend/src/domains/pickeat/restaurantExclude/components/restaurantTabList/restaurantList/RestaurantList.tsx
