+++ frontend/src/pages/pickeat/preferRestaurant/PreferRestaurant.tsx
@@ -1,4 +1,4 @@
-import PreferRestaurantList from '@domains/preferRestaurant/components/PreferRestaurantList';
+import PreferRestaurantList from '@domains/pickeat/preferRestaurant/components/PreferRestaurantList';
 
 import Button from '@components/actions/Button';
 import Arrow from '@components/assets/icons/Arrow';
@@ -6,7 +6,7 @@ import LoadingSpinner from '@components/assets/LoadingSpinner';
 import { HEADER_HEIGHT } from '@components/layouts/Header';
 
 import ErrorBoundary from '@domains/errorBoundary/ErrorBoundary';
-import { usePickeatStateChecker } from '@domains/matchResult/hooks/usePickeatEndCheck';
+import { usePickeatStateChecker } from '@domains/pickeat/matchResult/hooks/usePickeatEndCheck';
 
 import { restaurants } from '@apis/restaurants';
