+++ frontend/src/pages/pickeat/restaurantExclude/RestaurantExcludePage.tsx
@@ -1,11 +1,11 @@
-import PickeatEndModal from '@domains/matchResult/components/PickeatEndModal';
-import RestaurantExclude from '@domains/restaurantExclude/components/RestaurantExclude';
+import PickeatEndModal from '@domains/pickeat/matchResult/components/PickeatEndModal';
+import RestaurantExclude from '@domains/pickeat/restaurantExclude/components/RestaurantExclude';
 
 import LoadingSpinner from '@components/assets/LoadingSpinner';
 import { HEADER_HEIGHT } from '@components/layouts/Header';
 
 import ErrorBoundary from '@domains/errorBoundary/ErrorBoundary';
-import { usePickeatStateChecker } from '@domains/matchResult/hooks/usePickeatEndCheck';
+import { usePickeatStateChecker } from '@domains/pickeat/matchResult/hooks/usePickeatEndCheck';
 
 import { restaurants } from '@apis/restaurants';
